update main app
