+++ apps/backend-functions/src/data/types.ts
@@ -35,7 +35,7 @@ export {
   UserRole
 } from '@blockframes/permissions/types';
 export { PublicUser } from '@blockframes/auth/types';
-export { RequestDemoInformations } from '@blockframes/catalog/demo-request.model';
+export { RequestDemoInformations } from '@blockframes/apps/catalog/landing/demo-request.model';
 
 /**
  * Types used by the firebase backend.
