+++ apps/main/main/src/app/app-routing-module.ts
@@ -9,26 +9,54 @@ import { App } from '@blockframes/utils';
 import { createRoutes } from '@blockframes/utils/routes';
 
 /** Scaffold a dashboard like application routing for this application */
-const routes = createRoutes({
-  appName: 'main',
-  layout: LayoutComponent,
-  appsRoutes: [
-    {
-      path: '',
-      redirectTo: 'apps',
-      pathMatch: 'full'
-    },
-    {
-      path: 'apps',
-      component: AppGridComponent
-    },
-    {
-      path: App.mediaDelivering,
-      data: { app: App.mediaDelivering },
-      loadChildren: () => import('@blockframes/apps/delivery').then(m => m.DeliveryAppModule)
-    },
-  ]
-});
+// const routes = createRoutes({
+//   appName: 'main',
+//   appsRoutes: [
+//     {
+//       path: '',
+//       redirectTo: 'apps',
+//       pathMatch: 'full'
+//     },
+//     {
+//       path: 'apps',
+//       component: AppGridComponent
+//     },
+//     {
+//       path: 'catalog',
+//       loadChildren: () => import('@blockframes/apps/catalog/catalog.module').then(m => m.CatalogModule)
+//     },
+//     {
+//       path: App.mediaDelivering,
+//       data: { app: App.mediaDelivering },
+//       loadChildren: () => import('@blockframes/apps/delivery').then(m => m.DeliveryAppModule)
+//     },
+//     {
+//       path: App.biggerBoat,
+//       data: { app: App.biggerBoat },
+//       loadChildren: () =>
+//         import('@blockframes/apps/catalog/marketplace').then(m => m.CatalogMarketplaceAppModule)
+//     },
+//     {
+//       path: App.catalogDashboard,
+//       data: { app: App.catalogDashboard },
+//       loadChildren: () =>
+//         import('@blockframes/apps/catalog/dashboard').then(m => m.CatalogDashboardAppModule)
+//     }
+//   ]
+// });
+
+const routes = [{
+  path: '',
+  component: AppGridComponent
+},{
+  path: 'catalog',
+  loadChildren: () => import('@blockframes/apps/catalog/catalog.module').then(m => m.CatalogModule)
+},
+{
+  path: App.mediaDelivering,
+  data: { app: App.mediaDelivering },
+  loadChildren: () => import('@blockframes/apps/delivery').then(m => m.DeliveryAppModule)
+}]
 
 @NgModule({
   imports: [
