+++ libs/utils/src/lib/routes/create-routes.ts
@@ -4,26 +4,20 @@ import { PermissionsGuard, OrganizationGuard } from '@blockframes/organization';
 import { NotificationsGuard } from '@blockframes/notification';
 
 interface RouteOptions {
-  /** The Layout Component to render */
-  layout: any,
   /** The routes of the apps */
   appsRoutes: Routes,
   /** Name of the app to put in data of the route */
   appName: string,
-  /** The page to load by default */
-  rootPath?: string;
 }
 
-function root(LayoutComponent, children: Routes, rootPath: string, appName: string) {
+function root(children: Routes, appName: string) {
   return [
-    { path: '', redirectTo: rootPath, pathMatch: 'full' },
     {
       path: 'auth',
       loadChildren: () => import('@blockframes/auth/auth.module').then(m => m.AuthModule)
     },
     {
       path: 'c',
-      component: LayoutComponent,
       canActivate: [AuthGuard],
       canDeactivate: [AuthGuard],
       data: { app: appName },
@@ -59,7 +53,7 @@ function root(LayoutComponent, children: Routes, rootPath: string, appName: stri
 
 
 
-export function createRoutes({ layout, appsRoutes, appName, rootPath = 'c' }: RouteOptions) {
+export function createRoutes({ appsRoutes, appName }: RouteOptions) {
   const children = [
     ...appsRoutes,
     {
@@ -71,7 +65,7 @@ export function createRoutes({ layout, appsRoutes, appName, rootPath = 'c' }: Ro
       loadChildren: () => import('@blockframes/account').then(m => m.AccountModule)
     },
   ];
-  return root(layout, children, rootPath, appName);
+  return root(children, appName);
 }
